UTF8 fixes for 'std::filesystem' version of `IFileSystem`
On Linux/MacOS create a std::filesystem::path from a std::string assumes
the string is UTF-8 encoded.

But instead on Windows the string is assumed to be encoded according to
the current locale settings (this could be UTF-8, or it could be
something else. And ImGuiFileDialog should not make any assumptions
about the main application's locale).

Starting from C++20 it's possible to construct a std::filesystem::path
from a u8string (that's interpreted as a UTF-8 encoded string on all
platforms). But we can't use C++20 yet(?).

As a workaround, on windows, we construct std::filesystem::path from a
wstring. And to do that we first convert the std::string (which we
assume to be UTF-8 encoded) to wstring.

--

There's a similar problem with converting a std::filesystem::path back
to a std::string. On Linux/MacOS the std::filesystem::path::string()
method directly returns a UTF-8 encoded string. Unfortunately not on
Windows :(   So the workaround is to extract a wstring, and convert that
to a UTF-8 encoded string.

--

To reduce code duplication this patch adds two helper methods:
stringToPath() and pathToString(). So that the difference between
Linux/MacOS and Windows only needs to be handled once.

--

This patch also makes some fixes in CreateDirectoryIfNotExist().
* On windows this called std::filesystem::create_directory(), this
  creates only a single directory.
* But on Linux/MacOS this executed "mkdir -p ...", and that also creates
  all missing intermediate directories.
This patch changes both to use std::filesystem::create_directories().

There's also an emscripten version, I did not touch that one. Maybe that
one is also inconsistent? But then, does the combination of
std::filesystem and emscripten make sense??